Refactor: Extract `printFunctionParameters`
