+++ src/language-js/print/function-parameters.js
@@ -1,0 +1,166 @@
+"use strict";
+
+const {
+  getNextNonSpaceNonCommentCharacter,
+  isNextLineEmpty,
+} = require("../common/util");
+const { printDanglingComments } = require("../../main/comments");
+const {
+  builders: { concat, line, hardline, softline, group, indent, ifBreak },
+  utils: { removeLines },
+} = require("../../document");
+const {
+  getFunctionParameters,
+  iterateFunctionParametersPath,
+  isSimpleType,
+  isTestCall,
+  isTypeAnnotationAFunction,
+  isObjectTypePropertyAFunction,
+  hasRestParameter,
+  shouldPrintComma,
+} = require("../utils");
+const { locEnd } = require("../loc");
+const { shouldHugArguments } = require("./call-arguments");
+const { printFunctionTypeParameters } = require("./misc");
+
+function printFunctionParameters(
+  path,
+  print,
+  options,
+  expandArg,
+  printTypeParams
+) {
+  const functionNode = path.getValue();
+  const parameters = getFunctionParameters(functionNode);
+  const typeParams = printTypeParams
+    ? printFunctionTypeParameters(path, options, print)
+    : "";
+
+  if (parameters.length === 0) {
+    return concat([
+      typeParams,
+      "(",
+      printDanglingComments(
+        path,
+        options,
+        /* sameIndent */ true,
+        (comment) =>
+          getNextNonSpaceNonCommentCharacter(
+            options.originalText,
+            comment,
+            locEnd
+          ) === ")"
+      ),
+      ")",
+    ]);
+  }
+
+  const parent = path.getParentNode();
+  const isParametersInTestCall = isTestCall(parent);
+  const shouldHugParameters = shouldHugArguments(functionNode);
+  const shouldExpandParameters =
+    expandArg && !parameters.some((node) => node.comments);
+  const printed = [];
+  iterateFunctionParametersPath(path, (parameterPath, index) => {
+    const isLastParameter = index === parameters.length - 1;
+    if (isLastParameter && functionNode.rest) {
+      printed.push("...");
+    }
+    printed.push(parameterPath.call(print));
+    if (isLastParameter) {
+      return;
+    }
+    printed.push(",");
+    if (
+      isParametersInTestCall ||
+      shouldHugParameters ||
+      shouldExpandParameters
+    ) {
+      printed.push(" ");
+    } else if (
+      isNextLineEmpty(options.originalText, parameters[index], locEnd)
+    ) {
+      printed.push(hardline, hardline);
+    } else {
+      printed.push(line);
+    }
+  });
+
+  // If the parent is a call with the first/last argument expansion and this is the
+  // params of the first/last argument, we don't want the arguments to break and instead
+  // want the whole expression to be on a new line.
+  //
+  // Good:                 Bad:
+  //   verylongcall(         verylongcall((
+  //     (a, b) => {           a,
+  //     }                     b,
+  //   })                    ) => {
+  //                         })
+  if (shouldExpandParameters) {
+    return group(
+      concat([
+        removeLines(typeParams),
+        "(",
+        concat(printed.map(removeLines)),
+        ")",
+      ])
+    );
+  }
+
+  // Single object destructuring should hug
+  //
+  // function({
+  //   a,
+  //   b,
+  //   c
+  // }) {}
+  const hasNotParameterDecorator = parameters.every((node) => !node.decorators);
+  if (shouldHugParameters && hasNotParameterDecorator) {
+    return concat([typeParams, "(", concat(printed), ")"]);
+  }
+
+  // don't break in specs, eg; `it("should maintain parens around done even when long", (done) => {})`
+  if (isParametersInTestCall) {
+    return concat([typeParams, "(", concat(printed), ")"]);
+  }
+
+  const isFlowShorthandWithOneArg =
+    (isObjectTypePropertyAFunction(parent, options) ||
+      isTypeAnnotationAFunction(parent, options) ||
+      parent.type === "TypeAlias" ||
+      parent.type === "UnionTypeAnnotation" ||
+      parent.type === "TSUnionType" ||
+      parent.type === "IntersectionTypeAnnotation" ||
+      (parent.type === "FunctionTypeAnnotation" &&
+        parent.returnType === functionNode)) &&
+    parameters.length === 1 &&
+    parameters[0].name === null &&
+    // `type q = (this: string) => void;`
+    functionNode.this !== parameters[0] &&
+    parameters[0].typeAnnotation &&
+    functionNode.typeParameters === null &&
+    isSimpleType(parameters[0].typeAnnotation) &&
+    !functionNode.rest;
+
+  if (isFlowShorthandWithOneArg) {
+    if (options.arrowParens === "always") {
+      return concat(["(", concat(printed), ")"]);
+    }
+    return concat(printed);
+  }
+
+  return concat([
+    typeParams,
+    "(",
+    indent(concat([softline, concat(printed)])),
+    ifBreak(
+      !hasRestParameter(functionNode) && shouldPrintComma(options, "all")
+        ? ","
+        : ""
+    ),
+    softline,
+    ")",
+  ]);
+}
+
+module.export = { printFunctionParameters };
