+++ src/language-js/printer-estree.js
@@ -8,7 +8,6 @@ const assert = require("assert");
 // TODO(azz): anything that imports from main shouldn't be in a `language-*` dir.
 const comments = require("../main/comments");
 const {
-  getNextNonSpaceNonCommentCharacter,
   hasNewline,
   hasNewlineInRange,
   getLast,
@@ -19,11 +18,9 @@ const {
   hasNodeIgnoreComment,
   getIndentSize,
   getPreferredQuote,
-} = require("../common/util");
-const {
   isNextLineEmpty,
   getNextNonSpaceNonCommentCharacterIndex,
-} = require("../common/util-shared");
+} = require("../common/util");
 const {
   builders: {
     concat,
@@ -41,7 +38,7 @@ const {
     lineSuffixBoundary,
     addAlignmentToDoc,
   },
-  utils: { willBreak, isLineNext, isEmpty, removeLines, normalizeParts },
+  utils: { willBreak, isLineNext, isEmpty, normalizeParts },
   printer: { printDocToString },
 } = require("../document");
 const embed = require("./embed");
@@ -59,8 +56,6 @@ const {
   classPropMayCauseASIProblems,
   getFlowVariance,
   getFunctionParameters,
-  iterateFunctionParametersPath,
-  hasRestParameter,
   getLeftSidePathName,
   getParentExportDeclaration,
   getTypeScriptMappedTypeModifier,
@@ -104,7 +99,6 @@ const {
   isTestCall,
   isTheOnlyJSXElementInMarkdown,
   isTSXFile,
-  isTypeAnnotationAFunction,
   matchJsxWhitespaceRegex,
   needsHardlineAfterDanglingComment,
   rawText,
@@ -116,7 +110,10 @@ const {
 const { locStart, locEnd } = require("./loc");
 
 const printMemberChain = require("./print/member-chain");
-const printCallArguments = require("./print/call-arguments");
+const {
+  printCallArguments,
+  shouldHugArguments,
+} = require("./print/call-arguments");
 const {
   printOptionalToken,
   printFunctionTypeParameters,
@@ -125,6 +122,7 @@ const {
 } = require("./print/misc");
 const { printModuleSource, printModuleSpecifiers } = require("./print/module");
 const printTernaryOperator = require("./print/ternary");
+const { printFunctionParameters } = require("./print/function-parameters");
 
 const needsQuoteProps = new WeakMap();
 
@@ -3868,146 +3866,6 @@ function printTypeAnnotation(path, options, print) {
   ]);
 }
 
-function printFunctionParameters(
-  path,
-  print,
-  options,
-  expandArg,
-  printTypeParams
-) {
-  const functionNode = path.getValue();
-  const parameters = getFunctionParameters(functionNode);
-  const typeParams = printTypeParams
-    ? printFunctionTypeParameters(path, options, print)
-    : "";
-
-  if (parameters.length === 0) {
-    return concat([
-      typeParams,
-      "(",
-      comments.printDanglingComments(
-        path,
-        options,
-        /* sameIndent */ true,
-        (comment) =>
-          getNextNonSpaceNonCommentCharacter(
-            options.originalText,
-            comment,
-            locEnd
-          ) === ")"
-      ),
-      ")",
-    ]);
-  }
-
-  const parent = path.getParentNode();
-  const isParametersInTestCall = isTestCall(parent);
-  const shouldHugParameters = shouldHugArguments(functionNode);
-  const shouldExpandParameters =
-    expandArg && !parameters.some((node) => node.comments);
-  const printed = [];
-  iterateFunctionParametersPath(path, (parameterPath, index) => {
-    const isLastParameter = index === parameters.length - 1;
-    if (isLastParameter && functionNode.rest) {
-      printed.push("...");
-    }
-    printed.push(parameterPath.call(print));
-    if (isLastParameter) {
-      return;
-    }
-    printed.push(",");
-    if (
-      isParametersInTestCall ||
-      shouldHugParameters ||
-      shouldExpandParameters
-    ) {
-      printed.push(" ");
-    } else if (
-      isNextLineEmpty(options.originalText, parameters[index], locEnd)
-    ) {
-      printed.push(hardline, hardline);
-    } else {
-      printed.push(line);
-    }
-  });
-
-  // If the parent is a call with the first/last argument expansion and this is the
-  // params of the first/last argument, we don't want the arguments to break and instead
-  // want the whole expression to be on a new line.
-  //
-  // Good:                 Bad:
-  //   verylongcall(         verylongcall((
-  //     (a, b) => {           a,
-  //     }                     b,
-  //   })                    ) => {
-  //                         })
-  if (shouldExpandParameters) {
-    return group(
-      concat([
-        removeLines(typeParams),
-        "(",
-        concat(printed.map(removeLines)),
-        ")",
-      ])
-    );
-  }
-
-  // Single object destructuring should hug
-  //
-  // function({
-  //   a,
-  //   b,
-  //   c
-  // }) {}
-  const hasNotParameterDecorator = parameters.every((node) => !node.decorators);
-  if (shouldHugParameters && hasNotParameterDecorator) {
-    return concat([typeParams, "(", concat(printed), ")"]);
-  }
-
-  // don't break in specs, eg; `it("should maintain parens around done even when long", (done) => {})`
-  if (isParametersInTestCall) {
-    return concat([typeParams, "(", concat(printed), ")"]);
-  }
-
-  const isFlowShorthandWithOneArg =
-    (isObjectTypePropertyAFunction(parent, options) ||
-      isTypeAnnotationAFunction(parent, options) ||
-      parent.type === "TypeAlias" ||
-      parent.type === "UnionTypeAnnotation" ||
-      parent.type === "TSUnionType" ||
-      parent.type === "IntersectionTypeAnnotation" ||
-      (parent.type === "FunctionTypeAnnotation" &&
-        parent.returnType === functionNode)) &&
-    parameters.length === 1 &&
-    parameters[0].name === null &&
-    // `type q = (this: string) => void;`
-    functionNode.this !== parameters[0] &&
-    parameters[0].typeAnnotation &&
-    functionNode.typeParameters === null &&
-    isSimpleType(parameters[0].typeAnnotation) &&
-    !functionNode.rest;
-
-  if (isFlowShorthandWithOneArg) {
-    if (options.arrowParens === "always") {
-      return concat(["(", concat(printed), ")"]);
-    }
-    return concat(printed);
-  }
-
-  return concat([
-    typeParams,
-    "(",
-    indent(concat([softline, concat(printed)])),
-    ifBreak(
-      !hasRestParameter(functionNode) && shouldPrintComma(options, "all")
-        ? ","
-        : ""
-    ),
-    softline,
-    ")",
-  ]);
-}
-
 function shouldPrintParamsWithoutParens(path, options) {
   if (options.arrowParens === "always") {
     return false;
@@ -5112,37 +4970,6 @@ function shouldHugType(node) {
   }
 
   return false;
-}
-
-function shouldHugArguments(fun) {
-  if (!fun) {
-    return false;
-  }
-  const parameters = getFunctionParameters(fun);
-  if (parameters.length !== 1) {
-    return false;
-  }
-  const param = parameters[0];
-  return (
-    !param.comments &&
-    (param.type === "ObjectPattern" ||
-      param.type === "ArrayPattern" ||
-      (param.type === "Identifier" &&
-        param.typeAnnotation &&
-        (param.typeAnnotation.type === "TypeAnnotation" ||
-          param.typeAnnotation.type === "TSTypeAnnotation") &&
-        isObjectType(param.typeAnnotation.typeAnnotation)) ||
-      (param.type === "FunctionTypeParam" &&
-        isObjectType(param.typeAnnotation)) ||
-      (param.type === "AssignmentPattern" &&
-        (param.left.type === "ObjectPattern" ||
-          param.left.type === "ArrayPattern") &&
-        (param.right.type === "Identifier" ||
-          (param.right.type === "ObjectExpression" &&
-            param.right.properties.length === 0) ||
-          (param.right.type === "ArrayExpression" &&
-            param.right.elements.length === 0))))
-  );
 }
 
 function printArrayItems(path, options, printPath, print) {
